Core: rotate hue from the same matrix everywhere
